Fix bug of Integer2Long for split transactions

+++ app/src/main/java/com/money/manager/ex/nestedcategory/NestedCategoryListFragment.java
@@ -72,7 +72,7 @@ public class NestedCategoryListFragment
         implements LoaderManager.LoaderCallbacks<Cursor> {
 
     public String mAction = Intent.ACTION_EDIT;
-    public Long requestId;
+    public Integer requestId;
 
     private static final int ID_LOADER_CATEGORYSUB = 0;
